pyproject.toml
[project]
name = "photo-downloader"
version = "0.1.0"
description = ""
authors = [
    {name = "TyPaporotnyk",email = "daniil.svadkovsky@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "boto3 (>=1.42.40,<2.0.0)",
    "pyarrow (>=23.0.0,<24.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
